feat(nika-dap,nika-cli): the run journal + resume fold descend to the forensics plane
The run-cluster DESCENT wave, second leg (compute descends, render
stays). The journal format's owner is now ONE home: the writer sits
beside the walk that verifies it, the reader that recovers it, the
store that scans it, the retention that collects it.

To `nika-dap`:
- `journal` — the run-journal WRITE half from the run verb's sink.rs:
  `TraceFileSink` (the `.nika/traces/` NDJSON journal · lazy open ·
  owner-only · collision-safe naming), `JsonSink` (the `--json` lane),
  `Tee` (the fan-out the journal rides beside every primary surface),
  and the run-seal emission (`seal_journal` — the journal seals itself
  in the journal's home; `CARGO_PKG_VERSION` is the one workspace
  version on both sides of the old seam, sealed bytes unchanged). The
  genesis tag + chain hash are imported from `crate::chain` — the sink
  WRITES the same chain the walk verifies (one constant, one hash).
- `resume` — the ADR-099 fold from run/resume.rs: `fold_plan` +
  `apply_from` + `parse_answers` + `summary_line` + the `ResumeRequest`
  shape — a READER of the journal beside `recover`.

Stays in nika-cli (render, byte-identical UX): `FoldSink` (the live
TTY fold), `RenderMode`, the spinner rider, and `surface_trace` (the
`trace:` pointer's print lanes — now delegating the seal to
`seal_journal`). The `TRACE_DIR` constant dedups onto
`nika_dap::store::TRACE_DIR` (the store scan's home — one constant).

Consumers point at the new home directly (no re-export shims): run's
mod.rs, epilogue, child_runner and main.rs rewire to
`nika_dap::{journal,resume}`.

Zero behavior change: the code moves verbatim (docs and tests ride —
15 tests relocate, dap 138→153, cli 307→292), the journal bytes are
pinned by the same tests at the new home, and nika-display joins
dap's dev-deps for the demo event fixtures (test-only).
