+++ crates/nika-dap/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name          = "nika-dap"
-description   = "The Debug Adapter Protocol server (`nika dap`) — a READ-ONLY replay debugger over recorded run journals: breakpoints on task lines, stepping walks task settles, stepBack is free because the log is total (replay = re-render, never re-execute). Also the trace-forensics plane every reader shares: the chain walk, the tolerant recover, the OTLP projection, the reproduce comparison, the store scan and the retention policy. Layer L4 — split from nika-cli's verbs/dap 2026-07-09 (the crate-size cap · the nika-cap precedent), grown by the W0 trace descent the same day (compute descends, render stays)."
+description   = "The Debug Adapter Protocol server (`nika dap`) — a READ-ONLY replay debugger over recorded run journals: breakpoints on task lines, stepping walks task settles, stepBack is free because the log is total (replay = re-render, never re-execute). Also the trace-forensics plane every reader shares: the chain walk, the tolerant recover, the OTLP projection, the reproduce comparison, the store scan and the retention policy — and since 2026-07-22 the journal's WRITE half (the run-journal sink + the `--json` NDJSON lane + the Tee) and the ADR-099 resume fold, descended from nika-cli's run verb at the 15k wall (compute descends, render stays). Layer L4 — split from nika-cli's verbs/dap 2026-07-09 (the crate-size cap · the nika-cap precedent), grown by the W0 trace descent the same day."
 version.workspace      = true
 edition.workspace      = true
 license.workspace      = true
@@ -42,6 +42,7 @@ uuid           = { workspace = true }   # the RFC 3161 nonce entropy
 tempfile = { workspace = true }   # replay fixtures on disk
 proptest = { workspace = true }   # stats totality (quantile_h7 over arbitrary finite samples)
 nika-kernel-mock = { path = "../nika-kernel-mock", version = "0.105.0" }   # the anchor submit path's MockHttp seam
+nika-display = { path = "../nika-display", version = "0.105.0" }   # the journal tests' demo event fixtures (test-only)
 tokio    = { workspace = true }   # the submit path's test runtime (current-thread block_on)
 
 [package.metadata.lore]
